+++ lib/src/konsistArchitectureApiTest/kotlin/com/lemonappdev/konsist/architecture1/Architecture1Test.kt
@@ -1,8 +1,8 @@
 package com.lemonappdev.konsist.architecture1
 
 import com.lemonappdev.konsist.api.Konsist
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.architecture
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.assertArchitecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.architecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.assertArchitecture
 import com.lemonappdev.konsist.core.architecture.LayerImpl
 import org.junit.jupiter.api.Test
