+++ lib/src/konsistArchitectureApiTest/kotlin/com/lemonappdev/konsist/architecture4/Architecture4Test.kt
@@ -1,8 +1,8 @@
 package com.lemonappdev.konsist.architecture4
 
 import com.lemonappdev.konsist.api.Konsist
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.architecture
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.assertArchitecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.architecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.assertArchitecture
 import com.lemonappdev.konsist.core.architecture.LayerImpl
 import com.lemonappdev.konsist.core.exception.KoCheckFailedException
 import org.amshove.kluent.shouldThrow
