+++ lib/src/konsistArchitectureApiTest/kotlin/com/lemonappdev/konsist/architecture5/Architecture5Test.kt
@@ -1,7 +1,7 @@
 package com.lemonappdev.konsist.architecture5
 
 import com.lemonappdev.konsist.api.Konsist
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.assertArchitecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.assertArchitecture
 import com.lemonappdev.konsist.core.architecture.LayerImpl
 import com.lemonappdev.konsist.core.exception.KoPreconditionFailedException
 import org.amshove.kluent.shouldThrow
