+++ lib/src/konsistArchitectureApiTest/kotlin/com/lemonappdev/konsist/circulardependency3/CircularDependency3Test.kt
@@ -1,6 +1,6 @@
 package com.lemonappdev.konsist.circulardependency3
 
-import com.lemonappdev.konsist.api.architecture.KoArchitecture.architecture
+import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.architecture
 import com.lemonappdev.konsist.core.architecture.LayerImpl
 import com.lemonappdev.konsist.core.exception.KoPreconditionFailedException
 import org.amshove.kluent.shouldThrow
